+++ pkg/iter/sample_iterator_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cespare/xxhash"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -448,4 +449,55 @@ func Test_SampleSortIterator(t *testing.T) {
 			i++
 		}
 	})
+}
+
+func TestDedupeMergeSampleIterator(t *testing.T) {
+	it := NewMergeSampleIterator(context.Background(),
+		[]SampleIterator{
+			NewSeriesIterator(logproto.Series{
+				Labels: ``,
+				Samples: []logproto.Sample{
+					{
+						Timestamp: time.Unix(1, 0).UnixNano(),
+						Value:     1.,
+						Hash:      xxhash.Sum64String("1"),
+					},
+					{
+						Timestamp: time.Unix(1, 0).UnixNano(),
+						Value:     1.,
+						Hash:      xxhash.Sum64String("2"),
+					},
+				},
+				StreamHash: 0,
+			}),
+			NewSeriesIterator(logproto.Series{
+				Labels: ``,
+				Samples: []logproto.Sample{
+					{
+						Timestamp: time.Unix(1, 0).UnixNano(),
+						Value:     1.,
+						Hash:      xxhash.Sum64String("2"),
+					},
+					{
+						Timestamp: time.Unix(2, 0).UnixNano(),
+						Value:     1.,
+						Hash:      xxhash.Sum64String("3"),
+					},
+				},
+				StreamHash: 0,
+			}),
+		})
+
+	require.True(t, it.Next())
+	require.Equal(t, time.Unix(1, 0).UnixNano(), it.Sample().Timestamp)
+	require.Equal(t, 1., it.Sample().Value)
+	require.Equal(t, xxhash.Sum64String("1"), it.Sample().Hash)
+	require.True(t, it.Next())
+	require.Equal(t, time.Unix(1, 0).UnixNano(), it.Sample().Timestamp)
+	require.Equal(t, 1., it.Sample().Value)
+	require.Equal(t, xxhash.Sum64String("2"), it.Sample().Hash)
+	require.True(t, it.Next())
+	require.Equal(t, time.Unix(2, 0).UnixNano(), it.Sample().Timestamp)
+	require.Equal(t, 1., it.Sample().Value)
+	require.Equal(t, xxhash.Sum64String("3"), it.Sample().Hash)
 }