+++ pkg/storage/tsdb/bounds.go
@@ -1,9 +1,29 @@
 package tsdb
 
-import "github.com/prometheus/common/model"
+import (
+	"math"
+
+	"github.com/prometheus/common/model"
+)
 
 type Bounded interface {
 	Bounds() (model.Time, model.Time)
+}
+
+// InclusiveBounds will ensure the underlying Bounded implementation
+// is turned into [lower,upper] inclusivity.
+// Generally, we consider bounds to be `[lower,upper)` inclusive
+// This helper will account for integer overflow.
+// Because model.Time is millisecond-precise, but Loki uses nanosecond precision,
+// be careful usage can handle an extra millisecond being added.
+func inclusiveBounds(b Bounded) (model.Time, model.Time) {
+	lower, upper := b.Bounds()
+
+	if int64(upper) < math.MaxInt64 {
+		upper++
+	}
+
+	return lower, upper
 }
 
 type bounds struct {